Update core swig version to '4.0.0'.

+++ Examples/lua/arrays/example.c
@@ -13,10 +13,10 @@ void sort_int(int* arr, int len)
   qsort(arr, len, sizeof(int), compare_int);
 }
 
-// ditto doubles
+/* ditto doubles */
 int compare_double(const void * a, const void * b)
 {
-  return ( *(double*)a - *(double*)b );
+  return (int)( *(double*)a - *(double*)b );
 }
 
 void sort_double(double* arr, int len)
